ui: make better use of appbars and scaffolds for navigation
