--- app/src/main/java/com/goldenraven/devkitwallet/ui/components/AppBars.kt
@@ -1,51 +1,0 @@
-/*
- * Copyright 2021-2022 thunderbiscuit and contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
- */
-
-package com.goldenraven.devkitwallet.ui.components
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
-import com.goldenraven.devkitwallet.ui.Screen
-import com.goldenraven.devkitwallet.ui.theme.DevkitWalletColors
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun AwayFromHomeAppBar(navController: NavController, title: String) {
-    SmallTopAppBar(
-        title = { IntroAppTitle(title) },
-        colors = TopAppBarDefaults.smallTopAppBarColors(containerColor = DevkitWalletColors.primaryDark),
-        navigationIcon = {
-            IconButton(onClick = { navController.navigate(Screen.WalletScreen.route) }) {
-                Icon(
-                    imageVector = Icons.Filled.ArrowBack,
-                    contentDescription = "Go Back",
-                    tint = DevkitWalletColors.white
-                )
-            }
-        },
-        actions = { }
-    )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun IntroAppBar() {
-    SmallTopAppBar(
-        title = { IntroAppTitle("BDK Sample Wallet") },
-        colors = TopAppBarDefaults.smallTopAppBarColors(containerColor = DevkitWalletColors.primaryDark),
-        actions = { }
-    )
-}
-
-@Composable
-internal fun IntroAppTitle(title: String) {
-    Text(
-        text = title,
-        color = DevkitWalletColors.white,
-    )
-}
