+++ app/src/main/java/com/goldenraven/devkitwallet/ui/screens/wallet/WalletNavigation.kt
@@ -1,25 +1,24 @@
 /*
- * Copyright 2020-2022 thunderbiscuit and contributors.
+ * Copyright 2020-2023 thunderbiscuit and contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
  */
 
-package com.goldenraven.devkitwallet.ui.wallet
+package com.goldenraven.devkitwallet.ui.screens.wallet
 
 import androidx.compose.animation.AnimatedContentScope
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.material3.DrawerState
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import com.goldenraven.devkitwallet.ui.Screen
-import com.goldenraven.devkitwallet.ui.wallet.*
 import com.google.accompanist.navigation.animation.AnimatedNavHost
 import com.google.accompanist.navigation.animation.composable
 import com.google.accompanist.navigation.animation.rememberAnimatedNavController
 
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
-fun WalletNavigation(paddingValues: PaddingValues) {
+fun WalletNavigation(drawerState: DrawerState) {
     val navController: NavHostController = rememberAnimatedNavController()
     val animationDuration = 400
 
@@ -36,7 +35,7 @@ fun WalletNavigation(paddingValues: PaddingValues) {
             popEnterTransition = {
                 slideIntoContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
             },
-        ) { HomeScreen(navController, paddingValues) }
+        ) { WalletHomeScreen(navController, drawerState) }
 
         composable(
             route = Screen.ReceiveScreen.route,
@@ -52,7 +51,7 @@ fun WalletNavigation(paddingValues: PaddingValues) {
             popExitTransition = {
                 slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
             }
-        ) { ReceiveScreen(navController, paddingValues) }
+        ) { ReceiveScreen(navController) }
 
         composable(
             route = Screen.SendScreen.route,
@@ -68,7 +67,7 @@ fun WalletNavigation(paddingValues: PaddingValues) {
             popExitTransition = {
                 slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
             }
-        ) { SendScreen(navController, paddingValues) }
+        ) { SendScreen(navController) }
 
         composable(
             route = "${Screen.RBFScreen.route}/txid={txid}",
@@ -86,7 +85,7 @@ fun WalletNavigation(paddingValues: PaddingValues) {
             }
         ) { backStackEntry ->
             backStackEntry.arguments?.getString("txid")?.let {
-                RBFScreen(navController, paddingValues, backStackEntry.arguments?.getString("txid"))
+                RBFScreen(navController, backStackEntry.arguments?.getString("txid"))
             }
         }
 
@@ -104,7 +103,7 @@ fun WalletNavigation(paddingValues: PaddingValues) {
             popExitTransition = {
                 slideOutOfContainer(AnimatedContentScope.SlideDirection.End, animationSpec = tween(animationDuration))
             }
-        ) { TransactionsScreen(navController, paddingValues) }
+        ) { TransactionsScreen(navController) }
 
         composable(
             route = "${Screen.TransactionScreen.route}/txid={txid}",
@@ -122,7 +121,7 @@ fun WalletNavigation(paddingValues: PaddingValues) {
             }
         ) { backStackEntry ->
             backStackEntry.arguments?.getString("txid")?.let {
-                TransactionScreen(navController, paddingValues, backStackEntry.arguments?.getString("txid"))
+                TransactionScreen(navController, backStackEntry.arguments?.getString("txid"))
             }
         }
     }
